refactor(ui): rehome command router
